pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "groq>=1.0.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.5",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "tavily-python>=0.7.17",
    "torch>=2.9.1",
    "uvicorn>=0.40.0",
]
